+++ src/components/[guild]/EditButtonGroup/EditButtonGroup.tsx
@@ -8,10 +8,10 @@ import {
   useDisclosure,
 } from "@chakra-ui/react"
 import useGuild from "components/[guild]/hooks/useGuild"
-import { GearSix, PencilSimple, Plus } from "phosphor-react"
+import { GearSix, PaintBrush, PencilSimple, Plus } from "phosphor-react"
 import { useRef } from "react"
 import AddRoleDrawer from "./components/AddRoleDrawer"
-import CustomizationButton from "./components/CustomizationButton"
+import CustomizationModal from "./components/CustomizationModal"
 import EditGuildDrawer from "./components/EditGuildDrawer"
 
 const EditButtonGroup = (): JSX.Element => {
@@ -21,6 +21,12 @@ const EditButtonGroup = (): JSX.Element => {
     isOpen: isEditGuildDrawerOpen,
     onOpen: onEditGuildDrawerOpen,
     onClose: onEditGuildDrawerClose,
+  } = useDisclosure()
+
+  const {
+    isOpen: isCustomizationModalOpen,
+    onOpen: onCustomizationModalOpen,
+    onClose: onCustomizationModalClose,
   } = useDisclosure()
 
   const {
@@ -55,7 +61,14 @@ const EditButtonGroup = (): JSX.Element => {
             Edit guild
           </MenuItem>
 
-          <CustomizationButton />
+          <MenuItem
+            py="2"
+            cursor="pointer"
+            icon={<PaintBrush />}
+            onClick={onCustomizationModalOpen}
+          >
+            Customize appearance
+          </MenuItem>
 
           {guild?.platforms?.[0]?.platformType !== "DISCORD" && (
             <MenuItem
@@ -76,6 +89,14 @@ const EditButtonGroup = (): JSX.Element => {
           isOpen: isEditGuildDrawerOpen,
           onClose: onEditGuildDrawerClose,
           finalFocusRef: editGuildBtnRef,
+          children: null,
+        }}
+      />
+
+      <CustomizationModal
+        {...{
+          isOpen: isCustomizationModalOpen,
+          onClose: onCustomizationModalClose,
           children: null,
         }}
       />
