--- src/components/[guild]/EditButtonGroup/components/CustomizationButton/index.ts
@@ -1,3 +1,0 @@
-import CustomizationButton from "./CustomizationButton"
-
-export default CustomizationButton
