+++ src/components/[guild]/EditButtonGroup/components/CustomizationModal/components/RemoveBackgroundImage.tsx
@@ -1,6 +1,6 @@
 import { Button, Icon } from "@chakra-ui/react"
+import useEdit from "components/[guild]/hooks/useEdit"
 import { X } from "phosphor-react"
-import useEdit from "../hooks/useEdit"
 
 const RemoveBackgroundImage = () => {
   const { onSubmit, isLoading } = useEdit()
