+++ src/components/[guild]/EditButtonGroup/components/CustomizationModal/index.ts
@@ -1,0 +1,3 @@
+import CustomizationModal from "./CustomizationModal"
+
+export default CustomizationModal
