+++ src/components/[guild]/EditButtonGroup/components/EditGuildDrawer.tsx
@@ -18,7 +18,7 @@ import LogicPicker from "components/create-guild/LogicPicker"
 import NameAndIcon from "components/create-guild/NameAndIcon"
 import Requirements from "components/create-guild/Requirements"
 import DeleteGuildButton from "components/[guild]/edit/index/DeleteGuildButton"
-import useEdit from "components/[guild]/EditButtonGroup/components/CustomizationButton/hooks/useEdit"
+import useEdit from "components/[guild]/hooks/useEdit"
 import useGuild from "components/[guild]/hooks/useGuild"
 import usePersonalSign from "hooks/usePersonalSign"
 import useWarnIfUnsavedChanges from "hooks/useWarnIfUnsavedChanges"
